Extract log merge strategy
