+++ nidx/src/scheduler.rs
@@ -18,19 +18,18 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
 
+mod log_merge;
+mod merge_task;
 mod metrics_task;
-mod purge;
+mod purge_tasks;
 
-use crate::{
-    metadata::{MergeJob, SegmentId},
-    settings::MergeSettings,
-    NidxMetadata, Settings,
-};
+use crate::{metadata::MergeJob, settings::MergeSettings, NidxMetadata, Settings};
 use async_nats::jetstream::consumer::PullConsumer;
+use merge_task::MergeScheduler;
 use nidx_types::Seq;
 use object_store::DynObjectStore;
 // TODO: This should not be public but it's used in tests
-pub use purge::*;
+pub use purge_tasks::*;
 use std::{future::Future, sync::Arc, time::Duration};
 use tokio::{task::JoinSet, time::sleep};
 use tokio_util::sync::CancellationToken;
@@ -187,546 +186,4 @@ pub async fn retry_jobs(meta: &NidxMetadata) -> anyhow::Result<()> {
     }
 
     Ok(())
-}
-
-/// [`MergeScheduler`] implements a logarithmic merge strategy inspired from
-/// tantivy's log merge policy. The algorithm tries to merge segments with
-/// similar number of records.
-///
-/// It works by splitting up segments in different buckets depending on the
-/// amount of records they contain. Segments within a bucket are then scheduled
-/// to be merged together (if there are more than `min_number_of_segments`).
-///
-/// Buckets are dynamically computed depending on the segments. The biggest one
-/// marks the top bucket, and a configured bucket size marks the next ones.
-///
-/// To avoid merging segments too large or dividing in too many buckets for
-/// small segments, one can configure the top and bottom bucket ceilings.
-///
-/// Here's a graphical representation:
-///
-/// ```text
-///                              |   (too big to   |
-///                              |    be merged)   |
-///  `top_bucket_max_records` -> +-----------------+
-///                              |  (no segments)  |
-///                              +-----------------+ <- log2(biggest_segment.records)
-///                              |   top bucket    |
-///                              +-----------------+ <- log2(biggest_segment.records) - bucket_size_log
-///                              |                 |
-///                              +-----------------+ <- log2(biggest_segment.records) - 2 * bucket_size_log
-///                              |                 |
-///                              |       ...       |
-///                              |                 |
-///                          +-- +-----------------+
-///       `bucket_size_log` -+   |                 |
-///                          +-- +-----------------+
-///                              |                 |
-///                              +-----------------+ <- log2(biggest_segment.records) - n * bucket_size_log
-/// `bottom_bucket_threshold` -> +  -   -   -  -  -|
-///                              |  bottom bucket  |
-///                              |                 |
-///                              |   (too small    |
-///                              |    to split)    |
-/// ```
-///
-struct MergeScheduler {
-    /// Minimum number of segments needed to perform a merge for an index
-    min_number_of_segments: usize,
-
-    /// Max number of records for a segment to be in the top bucket, i.e.,
-    /// elegible for merge. Once a segment becomes bigger, it won't be merged
-    /// anymore
-    top_bucket_max_records: usize,
-
-    /// Max number of records for a segment to be considered in the bottom
-    /// bucket. Segments with fewer records won't be further splitted in buckets
-    bottom_bucket_threshold: usize,
-
-    /// Log value between buckets. Increasing this number implies more segment
-    /// sizes to be grouped in the same merge job.
-    bucket_size_log: f64,
-}
-
-impl Default for MergeScheduler {
-    fn default() -> Self {
-        Self {
-            min_number_of_segments: 4,
-            top_bucket_max_records: 100_000,
-            bottom_bucket_threshold: 1000,
-            bucket_size_log: 0.75,
-        }
-    }
-}
-
-impl MergeScheduler {
-    pub fn from_settings(settings: &MergeSettings) -> Self {
-        Self {
-            min_number_of_segments: settings.min_number_of_segments as usize,
-            top_bucket_max_records: settings.max_segment_size,
-            ..Default::default()
-        }
-    }
-
-    /// Enqueue merge jobs for segments older than `last_indexed_seq` that aren't
-    /// already scheduled for merge or marked to delete.
-    ///
-    /// Merging involves creation of a single segment from multiple ones, combining
-    /// their data and applying deletions. Merge jobs are executed in parallel (in
-    /// multiple workers) and while other segments are being indexed. This restricts
-    /// us to only merge segments whose sequences are less than the smaller sequence
-    /// being indexed.
-    ///
-    /// As an example, if sequences 100 and 102 are indexed but 101 is still being
-    /// indexed, we can only merge segments with sequence <= 100. Otherwise, if we
-    /// merge 100 and 102 (generating a new 102 segment) and segment 101 included
-    /// deletions for 100, we'll never apply them and we'll end in an inconsistent
-    /// state.
-    pub async fn schedule_merges(&self, meta: &NidxMetadata, last_indexed_seq: Seq) -> anyhow::Result<()> {
-        // TODO: Remove vector index split by metadata when removing tags
-        let merges = sqlx::query!(
-            r#"
-            SELECT
-                index_id,
-                array_agg(
-                    (segments.id, records)
-                    ORDER BY records DESC
-                ) AS "segments!: Vec<(SegmentId, i64)>"
-            FROM segments
-            JOIN indexes ON segments.index_id = indexes.id
-            WHERE
-                delete_at IS NULL AND merge_job_id IS NULL
-                AND seq <= $1
-                AND records <= $2
-            GROUP BY
-                index_id,
-                -- Only merge vector segments with same tags
-                CASE WHEN kind = 'vector' THEN index_metadata::text ELSE NULL END"#,
-            i64::from(last_indexed_seq),
-            self.top_bucket_max_records as i64,
-        )
-        .fetch_all(&meta.pool)
-        .await?;
-
-        for m in merges {
-            let mut buckets = vec![];
-            let mut current_bucket = vec![];
-            let mut current_max_size_log = f64::MAX;
-
-            for (segment_id, records) in m.segments {
-                let segment_size_log =
-                    f64::from(std::cmp::max(records as u32, self.bottom_bucket_threshold as u32)).log2();
-                if segment_size_log <= (current_max_size_log - self.bucket_size_log) {
-                    // traversed to next bucket, store current and continue
-                    buckets.push(current_bucket);
-                    current_bucket = vec![];
-                    current_max_size_log = segment_size_log;
-                }
-
-                current_bucket.push((segment_id, records));
-            }
-            buckets.push(current_bucket);
-
-            for segments in buckets {
-                // TODO: merge segments with too many deletions
-                if segments.len() >= self.min_number_of_segments {
-                    let mut sum_segments = 0;
-                    let mut to_merge = Vec::new();
-                    for (sid, records) in &segments {
-                        sum_segments += *records as usize;
-                        to_merge.push(*sid);
-                        if sum_segments > self.top_bucket_max_records {
-                            if to_merge.len() >= self.min_number_of_segments {
-                                debug!(?to_merge, "Scheduling merge job for bucket");
-                                MergeJob::create(meta, m.index_id.into(), &to_merge, last_indexed_seq).await?;
-                            }
-                            sum_segments = 0;
-                            to_merge.clear();
-                        }
-                    }
-
-                    if to_merge.len() >= self.min_number_of_segments {
-                        debug!(?to_merge, "Scheduling merge job for bucket");
-                        MergeJob::create(meta, m.index_id.into(), &to_merge, last_indexed_seq).await?;
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    mod merge_scheduling {
-        use std::collections::{HashMap, HashSet};
-
-        use nidx_types::Seq;
-        use nidx_vector::config::VectorConfig;
-        use serde_json::json;
-        use uuid::Uuid;
-
-        use super::*;
-
-        use crate::metadata::{Index, IndexConfig, IndexId, NidxMetadata, Segment, Shard};
-
-        fn merge_scheduler() -> MergeScheduler {
-            MergeScheduler::from_settings(&MergeSettings {
-                min_number_of_segments: 3,
-                ..Default::default()
-            })
-        }
-
-        async fn create_segments_with_num_records(
-            pool: &sqlx::PgPool,
-            index: &Index,
-            segment_records: &[i64],
-        ) -> anyhow::Result<Seq> {
-            let last_seq = sqlx::query_scalar!(r#"SELECT MAX(seq) from segments"#).fetch_one(pool).await?.unwrap_or(1);
-            let mut seq: i64 = last_seq;
-            for records in segment_records {
-                let segment =
-                    Segment::create(pool, index.id, Seq::from(seq), *records, serde_json::Value::Null).await?;
-                segment.mark_ready(pool, 512 * records).await?;
-                seq += 1;
-            }
-
-            Ok(Seq::from(seq - 1))
-        }
-
-        #[sqlx::test]
-        async fn test_log_merge_scheduling_not_enough_segments_merge(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-            let last_seq = create_segments_with_num_records(&meta.pool, &index, &[50; 3]).await?;
-
-            let scheduler = MergeScheduler::from_settings(&MergeSettings {
-                min_number_of_segments: 4,
-                ..Default::default()
-            });
-            scheduler.schedule_merges(&meta, last_seq).await?;
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert!(jobs.is_empty());
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_log_merge_scheduling_same_size_segments(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-            let last_seq = create_segments_with_num_records(&meta.pool, &index, &[50; 3]).await?;
-
-            let scheduler = MergeScheduler::from_settings(&MergeSettings {
-                min_number_of_segments: 3,
-                ..Default::default()
-            });
-            scheduler.schedule_merges(&meta, last_seq).await?;
-
-            // all segments have been scheduled for merge in a single job
-
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), 1);
-
-            let job = &jobs[0];
-            for segment in index.segments(&meta.pool).await? {
-                assert!(segment.merge_job_id.is_some_and(|id| id == job.id))
-            }
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_log_merge_scheduling_all_buckets(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-
-            let scheduler = MergeScheduler {
-                min_number_of_segments: 2,
-                top_bucket_max_records: 1000,
-                bottom_bucket_threshold: 50,
-                bucket_size_log: 1.0,
-            };
-
-            let last_seq = create_segments_with_num_records(
-                &meta.pool,
-                &index,
-                &[
-                    50,   // bottom bucket
-                    10,   // bottom bucket
-                    1000, // log2(1000) = ~9.97 -- will mark the top bucket
-                    63,   // bottom + 1
-                    124,  // bottom + 1
-                    62,   // first in bottom bucket
-                    1001, // exceeds the max segment size, won't appear
-                    20,   // bottom bucket
-                    125,  // top - 2
-                    51,   // just above bottom_bucket_threshold
-                    249,  // top - 2
-                    501,  // last element in top bucket
-                    500,  // just below top bucket
-                ],
-            )
-            .await?;
-            scheduler.schedule_merges(&meta, last_seq).await?;
-
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), 4);
-
-            // For test simplicity, we rely on jobs being created in
-            // top-to-bottom order. Feel free to change this is the algorithm
-            // changes
-            for segment in index.segments(&meta.pool).await? {
-                match segment.records {
-                    // > top_bucket_max_records
-                    1001 => {
-                        assert!(segment.merge_job_id.is_none());
-                    }
-                    // top bucket
-                    501 | 1000 => {
-                        assert!(segment.merge_job_id.is_some_and(|id| id == jobs[0].id))
-                    }
-                    // < min_number_of_segments
-                    500 => {
-                        assert!(segment.merge_job_id.is_none());
-                    }
-                    125 | 249 => {
-                        assert!(segment.merge_job_id.is_some_and(|id| id == jobs[1].id))
-                    }
-                    63 | 124 => {
-                        assert!(segment.merge_job_id.is_some_and(|id| id == jobs[2].id))
-                    }
-                    // bottom bucket
-                    10 | 20 | 50 | 51 | 62 => {
-                        assert!(segment.merge_job_id.is_some_and(|id| id == jobs[3].id))
-                    }
-                    r => {
-                        unreachable!(
-                            "All segment sizes should be handled in the match expression. Did you forgot about {r}?"
-                        )
-                    }
-                }
-            }
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_schedule_merges_for_shard_with_single_index(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-            let mut seq: i64 = 0;
-
-            for _ in 0..10 {
-                let segment =
-                    Segment::create(&meta.pool, index.id, Seq::from(seq), 50, serde_json::Value::Null).await?;
-                segment.mark_ready(&meta.pool, 1000).await?;
-                seq += 1;
-            }
-
-            let last_seq = Seq::from(seq - 1);
-
-            // creation of shards/indexes/segments don't trigger any merge job
-            assert!(MergeJob::take(&meta.pool).await?.is_none());
-
-            merge_scheduler().schedule_merges(&meta, last_seq).await?;
-
-            // one job has been scheduled for the index
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), 1);
-            assert_eq!(jobs[0].index_id, index.id);
-            assert_eq!(jobs[0].seq, last_seq);
-
-            for segment in index.segments(&meta.pool).await? {
-                assert!(segment.merge_job_id.is_some());
-                assert_eq!(segment.merge_job_id.unwrap(), jobs[0].id);
-            }
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_schedule_merges_for_shard_with_multiple_indexes(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-
-            let indexes = vec![
-                Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?,
-                Index::create(&meta.pool, shard.id, "english", VectorConfig::default().into()).await?,
-                Index::create(&meta.pool, shard.id, "fulltext", IndexConfig::new_text()).await?,
-                Index::create(&meta.pool, shard.id, "keyword", IndexConfig::new_paragraph()).await?,
-                Index::create(&meta.pool, shard.id, "relation", IndexConfig::new_relation()).await?,
-            ];
-            let mut seq: i64 = 0;
-
-            for _ in 0..10 {
-                for index in &indexes {
-                    let segment =
-                        Segment::create(&meta.pool, index.id, Seq::from(seq), 50, serde_json::Value::Null).await?;
-                    segment.mark_ready(&meta.pool, 1000).await?;
-                    seq += 1;
-                }
-            }
-
-            merge_scheduler().schedule_merges(&meta, Seq::from(seq)).await?;
-
-            // scheduled a job per index
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), indexes.len());
-            assert_eq!(
-                indexes.iter().map(|i| i.id).collect::<HashSet<_>>(),
-                jobs.iter().map(|j| j.index_id).collect::<HashSet<_>>(),
-            );
-
-            for job in &jobs {
-                assert_eq!(job.seq, Seq::from(seq));
-            }
-
-            // validate segments are marked with merge job id
-            let jobs_by_index: HashMap<IndexId, _> = jobs.iter().map(|job| (job.index_id, job)).collect();
-            for index in &indexes {
-                let (_, merge_job) = jobs_by_index.get_key_value(&index.id).unwrap();
-
-                let segments = index.segments(&meta.pool).await?;
-                assert_eq!(segments.len(), 10);
-
-                for segment in &segments {
-                    assert!(segment.merge_job_id.is_some());
-                    assert_eq!(segment.merge_job_id.unwrap(), merge_job.id);
-                }
-            }
-
-            Ok(())
-        }
-
-        async fn ongoing_indexing_scenario(pool: sqlx::PgPool) -> anyhow::Result<NidxMetadata> {
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-
-            for seq in [95, 98, 99, 100, 102i64] {
-                let segment =
-                    Segment::create(&meta.pool, index.id, Seq::from(seq), 50, serde_json::Value::Null).await?;
-                segment.mark_ready(&meta.pool, 1000).await?;
-            }
-
-            // 101 is still indexing
-            Segment::create(&meta.pool, index.id, Seq::from(101i64), 50, serde_json::Value::Null).await?;
-
-            Ok(meta)
-        }
-
-        #[sqlx::test]
-        async fn scheduling_with_smaller_than_existing_sequences(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = ongoing_indexing_scenario(pool).await?;
-
-            merge_scheduler().schedule_merges(&meta, Seq::from(50i64)).await?;
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert!(jobs.is_empty());
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn scheduling_with_ack_floor(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let meta = ongoing_indexing_scenario(pool).await?;
-            let ack_floor = Seq::from(100i64);
-
-            merge_scheduler().schedule_merges(&meta, ack_floor).await?;
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), 1);
-
-            let index = Index::get(&meta.pool, jobs[0].index_id).await?;
-            let segments = index.segments(&meta.pool).await?;
-
-            let mut segment_sequences = HashSet::new();
-            for segment in &segments {
-                if segment.seq <= ack_floor {
-                    // should be scheduled to merge
-                    assert!(segment.merge_job_id.is_some());
-                    assert_eq!(segment.merge_job_id.unwrap(), jobs[0].id);
-                    segment_sequences.insert(segment.seq);
-                } else {
-                    assert!(segment.merge_job_id.is_none());
-                }
-            }
-            let expected = [95, 98, 99, 100i64].into_iter().map(Seq::from).collect();
-            assert_eq!(segment_sequences, expected);
-
-            Ok(())
-        }
-
-        #[sqlx::test]
-        async fn test_schedule_merges_with_segment_tags(pool: sqlx::PgPool) -> anyhow::Result<()> {
-            let merge_scheduler = MergeScheduler::from_settings(&MergeSettings {
-                min_number_of_segments: 2,
-                ..Default::default()
-            });
-            let meta = NidxMetadata::new_with_pool(pool).await?;
-            let kbid = Uuid::new_v4();
-            let shard = Shard::create(&meta.pool, kbid).await?;
-
-            let index = Index::create(&meta.pool, shard.id, "multilingual", VectorConfig::default().into()).await?;
-
-            let hidden_count = 1;
-            let visible_count = 2;
-
-            let segment_hidden =
-                Segment::create(&meta.pool, index.id, 1i64.into(), hidden_count, json!({"tags": ["/q/h"]})).await?;
-            segment_hidden.mark_ready(&meta.pool, 1000).await?;
-
-            let segment_visible =
-                Segment::create(&meta.pool, index.id, 2i64.into(), visible_count, json!({"tags": []})).await?;
-            segment_visible.mark_ready(&meta.pool, 1000).await?;
-
-            // Cannot merge (two segments with different tags)
-            merge_scheduler.schedule_merges(&meta, 6i64.into()).await?;
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert!(jobs.is_empty());
-
-            let segment_hidden2 =
-                Segment::create(&meta.pool, index.id, 3i64.into(), hidden_count, json!({"tags": ["/q/h"]})).await?;
-            segment_hidden2.mark_ready(&meta.pool, 1000).await?;
-
-            let segment_visible2 =
-                Segment::create(&meta.pool, index.id, 4i64.into(), visible_count, json!({"tags": []})).await?;
-            segment_visible2.mark_ready(&meta.pool, 1000).await?;
-
-            // Can merge hidden and visible segments pair-wise
-            merge_scheduler.schedule_merges(&meta, 6i64.into()).await?;
-            let jobs = get_all_merge_jobs(&meta).await?;
-            assert_eq!(jobs.len(), 2);
-
-            for j in jobs {
-                let merged_segments = j.segments(&meta.pool).await?;
-                assert_eq!(merged_segments.len(), 2);
-                // We use different number of records to differentiante hidden/visible segments
-                // because the tags field is private outside the vectors index
-                assert_eq!(merged_segments[0].records, merged_segments[1].records);
-            }
-
-            Ok(())
-        }
-
-        async fn get_all_merge_jobs(meta: &NidxMetadata) -> anyhow::Result<Vec<MergeJob>> {
-            let mut jobs = vec![];
-            while let Some(job) = MergeJob::take(&meta.pool).await? {
-                jobs.push(job);
-            }
-            Ok(jobs)
-        }
-    }
 }
