+++ nidx/src/scheduler/purge_tasks.rs
